refactor: extract view components to files
